started refactoring tripsRepository into multiple subrepositories, currently done the technical part of expenses and notifications
