+++ app/src/main/java/com/github/se/wanderpals/model/repository/TripsRepository.kt
@@ -11,10 +11,8 @@ import com.github.se.wanderpals.model.data.Trip
 import com.github.se.wanderpals.model.data.TripNotification
 import com.github.se.wanderpals.model.data.User
 import com.github.se.wanderpals.model.firestoreData.FirestoreAnnouncement
-import com.github.se.wanderpals.model.firestoreData.FirestoreExpense
 import com.github.se.wanderpals.model.firestoreData.FirestoreStop
 import com.github.se.wanderpals.model.firestoreData.FirestoreSuggestion
-import com.github.se.wanderpals.model.firestoreData.FirestoreTripNotification
 import com.github.se.wanderpals.model.firestoreData.FirestoreUser
 import com.github.se.wanderpals.service.NotificationsManager
 import com.github.se.wanderpals.service.SessionManager
@@ -46,6 +44,14 @@ open class TripsRepository(
     private val dispatcher: CoroutineDispatcher // Inject dispatcher
 ) {
 
+  private lateinit var userRepository: UserRepository
+  private lateinit var tripRepository: TripRepository
+  private lateinit var expenseRepository: ExpenseRepository
+  private lateinit var announcementRepository: AnnouncementRepository
+  private lateinit var suggestionRepository: SuggestionRepository
+  private lateinit var stopRepository: StopRepository
+  private lateinit var notificationRepository: NotificationRepository
+
   private lateinit var firestore: FirebaseFirestore
   // Reference to the 'Users' collection in Firestore
   private lateinit var usersCollection: CollectionReference
@@ -54,7 +60,7 @@ open class TripsRepository(
 
   private lateinit var usernameCollection: CollectionReference
 
-  private val notificationsId = "Notifications"
+  // private val notificationsId = "Notifications"
   private val balancesId = "Balances"
 
   var isNetworkEnabled = false // by default we consider that this is false
@@ -69,6 +75,17 @@ open class TripsRepository(
   fun initFirestore(app: FirebaseApp) {
     Log.d("TripsRepository", "initFirestore(app): Initializing with specific FirebaseApp instance.")
     firestore = FirebaseFirestore.getInstance(app)
+
+    userRepository = UserRepository(firestore, dispatcher, uid, this)
+    tripRepository = TripRepository(firestore, dispatcher, uid, this)
+    expenseRepository = ExpenseRepository(firestore, dispatcher, uid, this)
+    announcementRepository = AnnouncementRepository(firestore, dispatcher, uid, this)
+    suggestionRepository = SuggestionRepository(firestore, dispatcher, uid, this)
+    stopRepository = StopRepository(firestore, dispatcher, uid, this)
+    notificationRepository = NotificationRepository(firestore, dispatcher, uid, this)
+    notificationRepository.init()
+    expenseRepository.init()
+
     usersCollection = firestore.collection(FirebaseCollections.USERS_TO_TRIPS_IDS.path)
     tripsCollection = firestore.collection(FirebaseCollections.TRIPS.path)
     usernameCollection = firestore.collection(FirebaseCollections.USERNAME_TO_EMAIL_COLLECTION.path)
@@ -82,6 +99,27 @@ open class TripsRepository(
   fun initFirestore() {
     Log.d("TripsRepository", "initFirestore: Initializing with default FirebaseApp.")
     firestore = FirebaseFirestore.getInstance()
+
+    userRepository = UserRepository(firestore, dispatcher, uid, this)
+    tripRepository = TripRepository(firestore, dispatcher, uid, this)
+    expenseRepository = ExpenseRepository(firestore, dispatcher, uid, this)
+    announcementRepository = AnnouncementRepository(firestore, dispatcher, uid, this)
+    suggestionRepository = SuggestionRepository(firestore, dispatcher, uid, this)
+    stopRepository = StopRepository(firestore, dispatcher, uid, this)
+    notificationRepository = NotificationRepository(firestore, dispatcher, uid, this)
+
+    /*
+         userRepository.init()
+         tripRepository.init()
+         expenseRepository.init()
+         announcementRepository.init()
+         suggestionRepository.init()
+         stopRepository.init()
+
+    */
+    notificationRepository.init()
+    expenseRepository.init()
+
     usersCollection = firestore.collection(FirebaseCollections.USERS_TO_TRIPS_IDS.path)
     tripsCollection = firestore.collection(FirebaseCollections.TRIPS.path)
     usernameCollection = firestore.collection(FirebaseCollections.USERNAME_TO_EMAIL_COLLECTION.path)
@@ -206,366 +244,45 @@ open class TripsRepository(
         }
       }
 
-  /**
-   * Retrieves the balance map for a specific trip from Firestore. This method queries the Firestore
-   * subcollection for balances associated with a given trip document identified by its unique trip
-   * ID. The balances are stored under a specific document, which is predefined by `balancesId`.
-   *
-   * @param tripId The unique identifier of the trip for which balances are to be fetched.
-   * @return A map of String to Double representing the balances for the trip. Returns an empty map
-   *   if the balances document is not found, if there are no balances in the document, or in case
-   *   of an error during data retrieval. Errors during the operation are logged and an empty map is
-   *   returned.
-   */
   open suspend fun getBalances(tripId: String): Map<String, Double> =
-      withContext(dispatcher) {
-        try {
-          val documentSnapshot =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_BALANCES_SUBCOLLECTION.path)
-                  .document(balancesId)
-                  .get(getSource())
-                  .await()
-          @Suppress("UNCHECKED_CAST")
-          (documentSnapshot.data as? Map<String, Double>) ?: emptyMap()
-        } catch (e: Exception) {
-          Log.e("TripsRepository", "getBalances: Error getting balance map for trip $tripId", e)
-          emptyMap()
-        }
-      }
+      expenseRepository.getBalances(tripId = tripId, source = getSource())
 
-  /**
-   * Sets or updates the balance map for a specific trip in Firestore. This method updates the
-   * Firestore subcollection for balances by setting a map of String to Double associated with a
-   * given trip document, identified by its unique trip ID. The balances are stored under a specific
-   * document predefined by `balancesId`.
-   *
-   * @param tripId The unique identifier of the trip for which balances are to be set.
-   * @param balancesMap The map of String to Double to be stored as balances for the trip.
-   * @return A Boolean value indicating the success or failure of the operation. Returns `true` if
-   *   the operation is successful, and `false` if there is an error during the setting operation.
-   *   Errors are logged.
-   */
   open suspend fun setBalances(tripId: String, balancesMap: Map<String, Double>): Boolean =
-      withContext(dispatcher) {
-        if (!checkNetworkIsValidAndLog()) {
-          return@withContext false
-        }
-        try {
-          val balanceDocument =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_BALANCES_SUBCOLLECTION.path)
-                  .document(balancesId)
+      expenseRepository.setBalances(tripId = tripId, balancesMap = balancesMap)
 
-          if (balancesMap.isEmpty()) {
-            balanceDocument.delete().await()
-          } else {
-            balanceDocument.set(balancesMap).await()
-          }
-          true
-        } catch (e: Exception) {
-          Log.e("TripsRepository", "setBalances: Error setting balance map for trip $tripId", e)
-          false
-        }
-      }
-
-  /**
-   * Retrieves a list of notifications for a specific trip from Firestore. This method queries the
-   * Firestore subcollection for notifications associated with a given trip document identified by
-   * its unique trip ID. The notifications are stored under a specific document, which is
-   * predefined.
-   *
-   * @param tripId The unique identifier of the trip for which notifications are to be fetched.
-   * @return A list of `TripNotification` objects, representing the notifications for the trip.
-   *   Returns an empty list if the notifications document is not found, if there are no
-   *   notifications in the document, or in case of an error during data retrieval. Errors during
-   *   the operation are logged and an empty list is returned.
-   */
   open suspend fun getNotificationList(tripId: String): List<TripNotification> =
-      withContext(dispatcher) {
-        try {
-          val documentSnapshot =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_NOTIFICATIONS_SUBCOLLECTION.path)
-                  .document(notificationsId)
-                  .get(getSource())
-                  .await()
-          // Attempt to retrieve the list using the correct type information
-          val data = documentSnapshot.data
-          @Suppress("UNCHECKED_CAST")
-          val notificationsList =
-              data?.get(notificationsId) as? List<Map<String, Any>> ?: emptyList()
+      notificationRepository.getNotificationList(tripId = tripId, source = getSource())
 
-          // Convert the list of map to list of FirestoreTripNotification
-          notificationsList.map { map ->
-            FirestoreTripNotification(
-                    title = map["title"] as String,
-                    route = map["route"] as String,
-                    timestamp = map["timestamp"] as String,
-                    navActionVariables = map["navActionVariables"] as String)
-                .toTripNotification()
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "getNotificationList: Error getting the Notification List for Trip $tripId.",
-              e)
-          emptyList()
-        }
-      }
-
-  /**
-   * Sets or updates a list of notifications for a specific trip in Firestore. This method updates
-   * the Firestore subcollection for notifications by setting a list of notifications associated
-   * with a given trip document, identified by its unique trip ID. The notifications are stored
-   * under a specific document predefined by `notificationsId`.
-   *
-   * @param tripId The unique identifier of the trip for which notifications are to be set.
-   * @param notifications The list of `TripNotification` objects to be stored as notifications for
-   *   the trip.
-   * @return A Boolean value indicating the success or failure of the operation. Returns `true` if
-   *   the operation is successful, and `false` if there is an error during the setting operation.
-   *   Errors are logged.
-   */
   open suspend fun setNotificationList(
       tripId: String,
       notifications: List<TripNotification>
   ): Boolean =
-      withContext(dispatcher) {
-        try {
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
+      if (checkNetworkIsValidAndLog())
+          notificationRepository.setNotificationList(
+              tripId = tripId, notifications = notifications, source = getSource())
+      else false
 
-          val notificationDocument =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_NOTIFICATIONS_SUBCOLLECTION.path)
-                  .document(notificationsId)
-
-          if (notifications.isEmpty()) {
-            notificationDocument.delete().await()
-            true
-          } else {
-            // Convert all TripNotifications to FirestoreTripNotifications
-            val firestoreNotificationList =
-                notifications.map { FirestoreTripNotification.fromTripNotification(it) }
-            // Set the notifications list in Firestore using the predefined notificationsId key
-            notificationDocument.set(mapOf(notificationsId to firestoreNotificationList)).await()
-            true
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "setNotificationList: Error setting the Notification List for Trip $tripId.",
-              e)
-          false
-        }
-      }
-
-  /**
-   * Retrieves a specific expense from a trip using the expense's unique identifier. This method
-   * queries the Firestore subcollection for trip expenses within a specific trip document.
-   *
-   * @param tripId The unique identifier of the trip.
-   * @param expenseId The unique identifier of the expense to be retrieved.
-   * @return An `Expense` object if found, or `null` if the expense is not found or if an error
-   *   occurs. The method logs an error and returns `null` in case of failure.
-   */
   open suspend fun getExpenseFromTrip(tripId: String, expenseId: String): Expense? =
-      withContext(dispatcher) {
-        try {
-          val documentSnapshot =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_EXPENSES_SUBCOLLECTION.path)
-                  .document(expenseId)
-                  .get(getSource())
-                  .await()
-          val firestoreExpense = documentSnapshot.toObject<FirestoreExpense>()
-          if (firestoreExpense != null) {
-            firestoreExpense.toExpense()
-          } else {
-            Log.e(
-                "TripsRepository",
-                "getExpenseFromTrip: Not found Expense $expenseId from trip $tripId.")
-            null
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "getExpenseFromTrip: Error getting Expense $expenseId from trip $tripId.",
-              e)
-          null // error
-        }
-      }
+      expenseRepository.getExpenseFromTrip(
+          tripId = tripId, expenseId = expenseId, source = getSource())
 
-  /**
-   * Fetches all expenses associated with a specific trip. This method retrieves an array of
-   * expenses by leveraging the `getExpenseFromTrip` method for each expense ID found in the trip
-   * document.
-   *
-   * @param tripId The unique identifier of the trip for which expenses are being retrieved.
-   * @return A list of `Expense` objects. Returns an empty list if the trip is not found or if an
-   *   error occurs during fetching. Errors are logged and an empty list is returned in these cases.
-   */
   open suspend fun getAllExpensesFromTrip(tripId: String): List<Expense> =
-      withContext(dispatcher) {
-        try {
-          val trip = getTrip(tripId)
+      expenseRepository.getAllExpensesFromTrip(tripId = tripId, source = getSource())
 
-          if (trip != null) {
-            val expenseIds = trip.expenses
-            expenseIds.mapNotNull { expenseId -> getExpenseFromTrip(tripId, expenseId) }
-          } else {
-            Log.e("TripsRepository", "getAllExpensesFromTrip: Trip not found with ID $tripId.")
-            emptyList()
-          }
-        } catch (e: Exception) {
-
-          Log.e(
-              "TripsRepository",
-              "getAllExpensesFromTrip: Error fetching Expenses to trip $tripId.",
-              e)
-          emptyList()
-        }
-      }
-
-  /**
-   * Adds a new expense to a specified trip. This method generates a unique identifier for the new
-   * expense, stores it in the Firestore, and updates the trip document to include this new expense
-   * ID.
-   *
-   * @param tripId The unique identifier of the trip to which the expense is being added.
-   * @param expense The `Expense` object to be added.
-   * @return The unique identifier of the newly added expense if the operation is successful, or an
-   *   empty string if it fails. The method logs the outcome of the operation.
-   */
   open suspend fun addExpenseToTrip(tripId: String, expense: Expense): String =
-      withContext(dispatcher) {
-        try {
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext ""
-          }
-          val uniqueID = UUID.randomUUID().toString()
-          val firebaseExpense =
-              FirestoreExpense.fromExpense(
-                  expense.copy(expenseId = uniqueID)) // we already know who creates the Expense
-          val expenseDocument =
-              tripsCollection
-                  .document(tripId)
-                  .collection(FirebaseCollections.TRIP_EXPENSES_SUBCOLLECTION.path)
-                  .document(uniqueID)
-          expenseDocument.set(firebaseExpense).await()
-          Log.d("TripsRepository", "addExpenseToTrip: Expense added successfully to trip $tripId.")
+      if (checkNetworkIsValidAndLog())
+          expenseRepository.addExpenseToTrip(tripId = tripId, expense = expense)
+      else ""
 
-          val trip = getTrip(tripId)
-          if (trip != null) {
-            // Add the new ExpenseId to the trip's Expense list and update the
-            // trip
-            val updatedExpensesList = trip.expenses + uniqueID
-            val updatedTrip = trip.copy(expenses = updatedExpensesList)
-            updateTrip(updatedTrip)
-            Log.d("TripsRepository", "addExpenseToTrip: Expense ID added to trip successfully.")
-
-            uniqueID // this function will return the new object ID if its successful
-          } else {
-
-            Log.e("TripsRepository", "addExpenseToTrip: Trip not found with ID $tripId.")
-
-            "" // And will return an empty string in case of failure
-          }
-        } catch (e: Exception) {
-          Log.e("TripsRepository", "addExpenseToTrip: Error adding Expense to trip $tripId.", e)
-          ""
-        }
-      }
-
-  /**
-   * Removes a specific expense from a trip. This method deletes the expense document from Firestore
-   * and updates the trip's expense list to exclude the removed expense ID.
-   *
-   * @param tripId The unique identifier of the trip.
-   * @param expenseId The unique identifier of the expense to be removed.
-   * @return `true` if the expense is successfully removed and the trip is updated; `false` if the
-   *   trip is not found or if any error occurs during the operation. Errors are logged.
-   */
   open suspend fun removeExpenseFromTrip(tripId: String, expenseId: String): Boolean =
-      withContext(dispatcher) {
-        try {
-          Log.d(
-              "TripsRepository",
-              "removeExpenseFromTrip: removing Expense $expenseId from trip $tripId")
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
-          tripsCollection
-              .document(tripId)
-              .collection(FirebaseCollections.TRIP_EXPENSES_SUBCOLLECTION.path)
-              .document(expenseId)
-              .delete()
-              .await()
+      if (checkNetworkIsValidAndLog())
+          expenseRepository.removeExpenseFromTrip(tripId = tripId, expenseId = expenseId)
+      else false
 
-          val trip = getTrip(tripId)
-          if (trip != null) {
-            val updatedExpensesList = trip.expenses.filterNot { it == expenseId }
-            val updatedTrip = trip.copy(expenses = updatedExpensesList)
-            updateTrip(updatedTrip)
-            Log.d(
-                "TripsRepository",
-                "removeExpenseFromTrip: Expense $expenseId remove and trip updated successfully.")
-            true
-          } else {
-            Log.e("TripsRepository", "removeExpenseFromTrip: Trip not found with ID $tripId.")
-            false
-          }
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "removeExpenseFromTrip: Error removing Expense $expenseId from trip $tripId.",
-              e)
-          false
-        }
-      }
-
-  /**
-   * Updates an existing expense within a trip. This method replaces the current expense document in
-   * Firestore with the updated expense data provided.
-   *
-   * @param tripId The unique identifier of the trip containing the expense.
-   * @param expense The updated `Expense` object.
-   * @return `true` if the expense is successfully updated; `false` if the trip is not found or an
-   *   error occurs. Errors are logged and the method returns `false` in these cases.
-   */
   open suspend fun updateExpenseInTrip(tripId: String, expense: Expense): Boolean =
-      withContext(dispatcher) {
-        try {
-          Log.d("TripsRepository", "updateExpenseInTrip: Updating a Expense in trip $tripId")
-          if (!checkNetworkIsValidAndLog()) {
-            return@withContext false
-          }
-          val firestoreExpense = FirestoreExpense.fromExpense(expense)
-          tripsCollection
-              .document(tripId)
-              .collection(FirebaseCollections.TRIP_EXPENSES_SUBCOLLECTION.path)
-              .document(firestoreExpense.expenseId)
-              .set(firestoreExpense)
-              .await()
-          Log.d(
-              "TripsRepository",
-              "updateExpenseInTrip: Trip's Expense updated successfully for ID $tripId.")
-          true
-        } catch (e: Exception) {
-          Log.e(
-              "TripsRepository",
-              "updateExpenseInTrip: Error updating Expense with ID ${expense.expenseId} in trip with ID $tripId",
-              e)
-          false
-        }
-      }
+      if (checkNetworkIsValidAndLog())
+          expenseRepository.updateExpenseInTrip(tripId = tripId, expense = expense)
+      else false
 
   /**
    * Retrieves a specific trip Announcement from a trip based on the Announcement's unique
